Fix unicode decode error if using umlauts in a filter.
user_filter_list is a list containing unicode values. The LDAP plugin needs utf-8 encoded strings and also the user-filter is encoded in utf-8. Thus we just convert all values to utf-8 to have a clean state.

Take care, this error will not raise an error because it will just be ignored through a try/except block in the upper scope.